Simplified docker image for rust
1. Made base image be stable rust

This avoids having a moving target of nightly build, which cause cache
misses, and we use fixed toolchain anyway.

2. simplified docker build to a single command

We had to build 3 images for the sole purpose of having some caching,
when docker itself performs caching.

3. Made docker image have an optional argument to build in debug

This allows to build the docker image in development without having
to rely on the developer's own environment. It makes it easy to
develop in integration.

4. Added docker build to workflow

So that we can test the build of the image.

5. Added instructions to README on how to build the image.

Another try.